requirements.txt
python-telegram-bot==13.15
bitget-python-sdk==1.0.2
requests